new2

+++ src/main/java/MODEL/book.java
@@ -11,8 +11,8 @@ public class book {
 	private String bauthor;
 	private String bpublish;
 	private Date bdate;
-	private BigDecimal bprice;
-	private BigDecimal bdiscount;
+	private double bprice;
+	private double bdiscount;
 	private String synopsis;
 	private long stock;
 	private String sontype;
@@ -54,16 +54,16 @@ public class book {
 	public void setBdate(Date bdate) {
 		this.bdate = bdate;
 	}
-	public BigDecimal getBprice() {
+	public double getBprice() {
 		return bprice;
 	}
-	public void setBprice(BigDecimal bprice) {
+	public void setBprice(double bprice) {
 		this.bprice = bprice;
 	}
-	public BigDecimal getBdiscount() {
+	public double getBdiscount() {
 		return bdiscount;
 	}
-	public void setBdiscount(BigDecimal bdiscount) {
+	public void setBdiscount(double bdiscount) {
 		this.bdiscount = bdiscount;
 	}
 	public String getSynopsis() {
@@ -90,8 +90,8 @@ public class book {
 				+ bpublish + ", bdate=" + bdate + ", bprice=" + bprice + ", bdiscount=" + bdiscount + ", synopsis="
 				+ synopsis + ", stock=" + stock + ", sontype=" + sontype + "]";
 	}
-	public book(String iSBN, String img, String bname, String bauthor, String bpublish, Date bdate, BigDecimal bprice,
-			BigDecimal bdiscount, String synopsis, long stock, String sontype) {
+	public book(String iSBN, String img, String bname, String bauthor, String bpublish, Date bdate, double bprice,
+			double bdiscount, String synopsis, long stock, String sontype) {
 		super();
 		ISBN = iSBN;
 		this.img = img;
